+++ src/Sdk.php
@@ -3,6 +3,9 @@
 namespace Sentry;
 
 use Sentry\Breadcrumbs\Breadcrumb;
+use Sentry\Integration\ErrorHandlerIntegration;
+use Sentry\Integration\ExceptionIntegration;
+use Sentry\Integration\RequestIntegration;
 use Sentry\State\Hub;
 
 /**
@@ -12,7 +15,13 @@ use Sentry\State\Hub;
  */
 function init(array $options = []): void
 {
-    Hub::setCurrent(new Hub(ClientBuilder::create($options)->getClient()));
+    $options = new Options($options);
+    $client = new Client($options, [
+        new ErrorHandlerIntegration(),
+        new RequestIntegration(),
+        new ExceptionIntegration($options),
+    ]);
+    Hub::setCurrent(new Hub($client));
 }
 
 /**
